Use English as the default language
To avoid a situation where a fresh installation from this template is a mix of several languages, I have modified the template to include content only in English and to work with the English locale by default. I have also changed all references to "WP-palvelu" to "Seravo" and added links to currently relevant documentation.